Move intermediate functionality to entrypoints
